Add mandatory fields for catalog description in 5-catalog-and-complet… (#350)
* Add mandatory fields for catalog description in 5-catalog-and-complete-description.nb.mdx

Adds mandatory fields and an additional contact point for the catalog description in the  turtle example. 

When this addition is approved, I will make the same changes in the english and nynorsk version of the page.

* Fix: add correct id for contact points in example on 5-catalog-and-complete-description.nb.mdx

After the previous change, the contact point IDs in the example was no longer correct. This fix resolves this issue.

* Update 5-catalog-and-complete-description.nb.mdx

Fixes include: 

Replace blank nodes with URI. 
Add "mailto:" before email in contact point
Add languagetag behind name of contact point.

* Update 5-catalog-and-complete-description.en.mdx

Updates the example of complete example with catalog, dataset and distributions, according to the changes that has been done on the page in norwegian bokmål.

* Update 5-catalog-and-complete-description.nn.mdx

Update example according to changes that has been done on the page in norwegian bokmål.

* Fix:remove duplicate info 5-catalog-and-complete-description.en.mdx

Remove duplicate info

* FIx add correct provenance and add contact page to contact point on 5-catalog-and-complete-description.nb.mdx

This alters the provenance property, in order for it to be compliant to dcat-ap-no v.3

I have also added contact page to the contact point "Informasjonsforvaltning".

* Fix provenance property and add contact page to contact point on 5-catalog-and-complete-description.en.mdx 

Fixes provenance and adds contact page to contact poin "informasjonsforvaltning".

* Fix provenance and add contact page for contact point on 5-catalog-and-complete-description.nn.mdx

Fixes provenance, in order to make it compatible to dcat-ap-no v3. Adds contact page to contact point.

* Change: remove api-description, add api to distribution 5-catalog-and-complete-description.nb.mdx

Changes made: 
Removed API description from the turtle example. 
Added a reference to an API to distribution. 
Changed the sentence above the example, to be consistent with the changes made in the description.

* Fix rdf syntax in example on 5-catalog-and-complete-description.nb.mdx

Fix rdf syntax

* Change: example of complete description on 5-catalog-and-complete-description.en.mdx

Changes made: 
Removed dataservice from example of complete description. 
Added a link access service to distribution in example of complete description. 
Updated text above example, in order for it to be correct after the changes I made in the example.

* Change: update example of complete description on 5-catalog-and-complete-description.nn.mdx

Changes made: 
Removed API description from the example
Added a link to API to distribution in the example
Changed the text above the example, in order for it to be correct according to the changes I made in the example.

* Fix: remove reference to Eurovoc from example on 5-catalog-and-complete-description.nb.mdx

Fix: 
Eurovoc is no longer a recommended vocabulary for the property dcat:theme. I have removed it from the example.

* Fix: Remove reference to Eurovoc from example on 5-catalog-and-complete-description.en.mdx

Fix: 
Eurovoc is no longer a recommended vocabulary for the property dcat:theme. I have removed it from the example.

* Fix: Remove reference to eurovoc from example on 5-catalog-and-complete-description.nn.mdx

Fix: 
Eurovoc is no longer a recommended vocabulary for the property dcat:theme. I have removed it from the example.

* Fix Add link to distribution from dataset in example on 5-catalog-and-complete-description.nb.mdx

* Fix Add link to distribution from dataset in example on 5-catalog-and-complete-description.en.mdx

* Fix Add link to distribution from dataset in example on 5-catalog-and-complete-description.nn.mdx

* Fix: remove dct:spatial from example from 5-catalog-and-complete-description.nb.mdx

Removes dct:spatial from the example, because we don't support the code lists that are mandatory for this property yet.

* Fix: remove dct:spatial from example on 5-catalog-and-complete-description.en.mdx

Removes dct:spatial from the example, because we don't support the code lists that are mandatory for this property yet.

* Fix: remove dct:spatial from example on 5-catalog-and-complete-description.nn.mdx

Removes dct:spatial from the example, because we don't support the code lists that are mandatory for this property yet.