Chore: Update about.nb.mdx (#356)
* Chore: Update about.nb.mdx

Changes are made after input from our legal resource. We have focused on using the correct terms and downplaying private organizations.

* Change: Edit text on about.nb.mdx

Changes are made in order to clarify the mission of data.norge.no and limits of use.

* Fix: add bulletpoint that was missing on about.nb.mdx

Added a bullet point under "vil du vite mer?" to align this description with the one in nynorsk.

* Change: Align text to changes made on the page in nb on about.nn.mdx

Updatees the text, in order for it to be aligned with the changes we made on about.nb.mdx.

* Change: Align text to new text on the nb page on about.en.mdx

Changes are made to align the english page to the changes we made on about.nb.mdx

* Fix: add space about.en.mdx

Adds space between hashtags and heading

* Fix page title on about.en.mdx

Translated page title to english

* Fix: add public sector in ingress on Update about.en.mdx

Adds public sector to align the text with the norwegian version.

* Fix: add offentleg sektor to ingress on Update about.nn.mdx

Adds offentleg sektor to ingress to align the text with the bokmål version.

* Fix typo on about.en.mdx

* Fix: typo about.nn.mdx

* Fix typo about.en.mdx

* Fix better text flow and typos about.en.mdx

Fixed some sentences to make the text flow better. Also fixed some typos.

* Fix: update link text Update about.en.mdx

* Fix: small text edit Update about.nb.mdx

* Fix: small text edit Update about.nn.mdx